Refactorization of Integrators function
- Add documentation for GaussianPrimitive and ContractedOrbital.
- Move overlap_integral, laplacian_integral, electron_nucleus_integral and electron_electron_integral next to the associated orbital definition.
