+++ include/BasisSet/contracted_orbital.hpp
@@ -2,108 +2,143 @@
 
 #include <vector>
 
+#include "BasisSet/contracted_orbital.interface.hpp"
+#include "BasisSet/slater_primitive.hpp"
+#include "BasisSet/gaussian_primitive.hpp"
 #include "concepts.hpp"
-#include "Integrators/overlap_primitive.hpp"
+
 
 template <DerivedFromOrbital PrimitiveType>
-class ContractedOrbital final : public Orbital {
-public:
-    ContractedOrbital() = default;
+ContractedOrbital<PrimitiveType>::ContractedOrbital(std::vector<double> &coefficients, std::vector<PrimitiveType> &primitives) : m_coefficients(coefficients), m_primitives(primitives) {
+    if (coefficients.size() != primitives.size()) {
+        throw std::length_error("ContractedOrbital: The number of coefficients and primitives must be equal.");
+    }
 
-    /**
-     * @param size Number of primitives for the linear combination.
-     */
-    ContractedOrbital(size_t size) : m_coefficients(size), m_primitives(size) {}
+    update_normalization();
+}
 
-    /**
-     * @param coefficients Weights for each primitive in the linear combination.
-     * @param primitives Primitive basis functions.
-     * @throws std::length_error if the number of coefficients and primitives are not equal.
-     */
-    ContractedOrbital(std::vector<double> &coefficients, std::vector<PrimitiveType> &primitives) : m_coefficients(coefficients), m_primitives(primitives) {
-        if (coefficients.size() != primitives.size()) {
-            throw std::length_error("ContractedOrbital: The number of coefficients and primitives must be equal.");
+template <DerivedFromOrbital PrimitiveType>
+ContractedOrbital<PrimitiveType>::ContractedOrbital(const ContractedOrbital<PrimitiveType> &other) : m_coefficients(other.m_coefficients), m_primitives(other.m_primitives) {
+    m_normalization_constant = other.m_normalization_constant;
+}
+
+template <DerivedFromOrbital PrimitiveType>
+ContractedOrbital<PrimitiveType>::ContractedOrbital(const ContractedOrbital<PrimitiveType> &&other) : m_coefficients(std::move(other.m_coefficients)), m_primitives(std::move(other.m_primitives)) {
+    m_normalization_constant = other.m_normalization_constant;
+}
+
+template <DerivedFromOrbital PrimitiveType>
+void ContractedOrbital<PrimitiveType>::reserve(std::size_t size) {
+    m_coefficients.reserve(size);
+    m_primitives.reserve(size);
+}
+
+template <DerivedFromOrbital PrimitiveType>
+void ContractedOrbital<PrimitiveType>::add_primitive(double coefficient, const PrimitiveType &primitive) {
+    m_coefficients.push_back(coefficient);
+    m_primitives.push_back(primitive);
+    update_normalization();
+}
+
+template <DerivedFromOrbital PrimitiveType>
+void ContractedOrbital<PrimitiveType>::update_normalization() {
+    m_normalization_constant = 0.0;
+
+    for (size_t i = 0; i < m_primitives.size(); i++) {
+        for (size_t j = i; j < m_primitives.size(); j++) {
+            if (i == j) {
+                m_normalization_constant += m_coefficients[i] * m_coefficients[j] * overlap_integral(m_primitives[i], m_primitives[j]);
+                continue;
+            }
+
+            m_normalization_constant += 2 * m_coefficients[i] * m_coefficients[j] * overlap_integral(m_primitives[i], m_primitives[j]);
         }
-
-        update_normalization();
     }
 
-    ContractedOrbital(const ContractedOrbital<PrimitiveType> &other) : m_coefficients(other.m_coefficients), m_primitives(other.m_primitives) {
-        m_normalization_constant = other.m_normalization_constant;
+    m_normalization_constant = 1.0 / std::sqrt(m_normalization_constant);
+}
+
+template <DerivedFromOrbital PrimitiveType>
+inline size_t ContractedOrbital<PrimitiveType>::size() const {
+    return m_primitives.size();
+}
+
+template <DerivedFromOrbital PrimitiveType>
+inline const PrimitiveType& ContractedOrbital<PrimitiveType>::get_primitive(int i) const {
+    return m_primitives[i];
+}
+
+template <DerivedFromOrbital PrimitiveType>
+inline double ContractedOrbital<PrimitiveType>::get_coefficient(int i) const {
+    return m_coefficients[i];
+}
+
+
+// ===============================================================================================
+// ============================ Matrix element for Contracted Orbital ============================
+// ===============================================================================================
+
+template <DerivedFromOrbital PrimitiveType>
+double overlap_integral(const ContractedOrbital<PrimitiveType>& orbital1, const ContractedOrbital<PrimitiveType>& orbital2) {
+    double integral = 0.0;
+
+    for (size_t i = 0; i < orbital1.size(); i++) {
+        for (size_t j = 0; j < orbital2.size(); j++) {
+            integral += orbital1.get_coefficient(i) * orbital2.get_coefficient(j) *
+                        overlap_integral(orbital1.get_primitive(i), orbital2.get_primitive(j));
+        }
     }
 
-    ContractedOrbital(const ContractedOrbital<PrimitiveType> &&other) : m_coefficients(std::move(other.m_coefficients)), m_primitives(std::move(other.m_primitives)) {
-        m_normalization_constant = other.m_normalization_constant;
+    return integral * orbital1.normalization() * orbital2.normalization();
+}
+
+template <DerivedFromOrbital PrimitiveType>
+double laplacian_integral(const ContractedOrbital<PrimitiveType>& orbital1,
+                        const ContractedOrbital<PrimitiveType>& orbital2) {
+    double integral = 0.0;
+
+    for (size_t i = 0; i < orbital1.size(); i++) {
+        for (size_t j = 0; j < orbital2.size(); j++) {
+            integral += orbital1.get_coefficient(i) * orbital2.get_coefficient(j) *
+                        laplacian_integral(orbital1.get_primitive(i), orbital2.get_primitive(j));
+        }
     }
 
-    /**
-     * @brief Allocate memory for primitives and coefficients.
-     *
-     * @param size Number of primitives for the linear combination.
-     */
-    void reserve(std::size_t size) {
-        m_coefficients.reserve(size);
-        m_primitives.reserve(size);
+    return integral * orbital1.normalization() * orbital2.normalization();
+}
+
+template <DerivedFromOrbital PrimitiveType>
+double electron_nucleus_integral(const ContractedOrbital<PrimitiveType>& orbital1,
+                                const ContractedOrbital<PrimitiveType>& orbital2,
+                                const Eigen::Vector3d& nucleus_position) {
+    double integral = 0.0;
+    for (size_t i = 0; i < orbital1.size(); i++) {
+        for (size_t j = 0; j < orbital2.size(); j++) {
+            integral += orbital1.get_coefficient(i) * orbital2.get_coefficient(j) *
+                        electron_nucleus_integral(orbital1.get_primitive(i), orbital2.get_primitive(j), nucleus_position);
+        }
     }
 
-    void add_primitive(double coefficient, const PrimitiveType &primitive) {
-        m_coefficients.push_back(coefficient);
-        m_primitives.push_back(primitive);
-        update_normalization();
-    }
+    return integral * orbital1.normalization() * orbital2.normalization();
+}
 
-    template <typename... Args>
-    void add_primitive(double coefficient, Args&&... args) {
-        m_coefficients.push_back(coefficient);
-        m_primitives.emplace_back(std::forward<Args>(args)...);
-        update_normalization();
-    }
+template <DerivedFromOrbital PrimitiveType>
+double electron_electron_integral(const ContractedOrbital<PrimitiveType>& orbital1,
+                                const ContractedOrbital<PrimitiveType>& orbital2,
+                                const ContractedOrbital<PrimitiveType>& orbital3,
+                                const ContractedOrbital<PrimitiveType>& orbital4) {
+    double integral = 0.0;
 
-
-    void update_normalization() {
-        m_normalization_constant = 0.0;
-
-        for (size_t i = 0; i < m_primitives.size(); i++) {
-            for (size_t j = i; j < m_primitives.size(); j++) {
-                if (i == j) {
-                    m_normalization_constant += m_coefficients[i] * m_coefficients[j] * overlap_integral(m_primitives[i], m_primitives[j]);
-                    continue;
+    for (size_t i = 0; i < orbital1.size(); i++) {
+        for (size_t j = 0; j < orbital2.size(); j++) {
+            for (size_t k = 0; k < orbital3.size(); k++) {
+                for (size_t l = 0; l < orbital4.size(); l++) {
+                    integral += orbital1.get_coefficient(i) * orbital2.get_coefficient(j) * orbital3.get_coefficient(k) * orbital4.get_coefficient(l) *
+                                electron_electron_integral(orbital1.get_primitive(i), orbital2.get_primitive(j), orbital3.get_primitive(k), orbital4.get_primitive(l));
                 }
-
-                m_normalization_constant += 2 * m_coefficients[i] * m_coefficients[j] * overlap_integral(m_primitives[i], m_primitives[j]);
             }
         }
-
-        m_normalization_constant = 1.0 / std::sqrt(m_normalization_constant);
     }
 
-    inline size_t size() const {
-        return m_primitives.size();
-    }
-
-    inline const PrimitiveType& get_primitive(int i) const {
-        return m_primitives[i];
-    }
-
-    inline double get_coefficient(int i) const {
-        return m_coefficients[i];
-    }
-
-    void set_position(Eigen::Vector3d position) override {
-        this->Orbital::set_position(position);
-
-        for (PrimitiveType &primitive : m_primitives) {
-            primitive.set_position(position);
-        }
-    }
-
-private:
-    std::vector<double> m_coefficients;
-    std::vector<PrimitiveType> m_primitives;
-};
-
-extern template class ContractedOrbital<GaussianPrimitive>;
-extern template class ContractedOrbital<SlaterPrimitive>;
-
-typedef ContractedOrbital<GaussianPrimitive> ContractedGaussian;
-typedef ContractedOrbital<SlaterPrimitive> ContractedSlater;
+    return integral * orbital1.normalization() * orbital2.normalization() * orbital3.normalization() * orbital4.normalization();
+}
